bluprints for auth and admin
